🚀 feat: #13 パネルリストコンポーネントを、プレイモード時のみ表示するコンポーネントとで分割
